% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_EMV}
\alias{get_EMV}
\title{Get US Equity Market Volatility Index (TPU) data}
\usage{
get_EMV(all = T)
}
\arguments{
\item{all}{logical. if TRUE return all EMV categories.}
}
\value{
return an xts data object
}
\description{
Get US Equity Market Volatility Index (TPU) data
}
\examples{
emv_data <- get_EMV(all = FALSE)
dygraph(emv_data)
}
\references{
\url{https://www.policyuncertainty.com/EMV_monthly.html}
}
\seealso{
\code{\link{xts}}
}
