% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_epu}
\alias{get_epu}
\title{Get EPU indices}
\usage{
get_epu(region = "all")
}
\arguments{
\item{region}{a character indicating the region of the EPU you want. The defaul
is "all" regions. The region names has to be one of the options in the EPU country list}
}
\value{
return an xts data object containing the EPU for the chosen region
}
\description{
Takes in the region name
}
\seealso{
xts
}
