% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_FSI}
\alias{get_FSI}
\title{Get Financial Stress Indicator (FSI) data}
\usage{
get_FSI(freq = "monthly")
}
\arguments{
\item{freq}{either "monthly" or "quarterly"}
}
\value{
return an xts data object
}
\description{
Get Financial Stress Indicator (FSI) data
}
\details{
The FSI runs from 1889 to 2016 and is available at monthly and quarterly frequencies.
Index is constructed from the titles of articles published in five U.S.
newspapers: the Boston Globe, Chicago Tribune, Los Angeles Times, Wall Street Journal and Washington Post.
}
\examples{
\dontrun{
# for monthly FSI
fsi_mon <- get_FSI()
plot(fsi_mon)
fsi_quar <- get_FSI("quarterly")
plot(fsi_quar)
}
}
\references{
\url{https://www.policyuncertainty.com/financial_stress.html}
}
\seealso{
\code{\link{xts}}
}
