% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_OMI}
\alias{get_OMI}
\title{Get Oxford-Man Institute (OMI) data}
\usage{
get_OMI(index = "AEX")
}
\arguments{
\item{index, }{a character string of the index name needed.}
}
\value{
an xts data object
}
\description{
Realized volatility data from the Oxford-Man Institute of Quantitative Finance website
}
\examples{
\dontrun{
## it take about 30-60 seconds to run this example
## Its size is about 15Mb in zip
aex_data <- get_OMI("AEX")
str(aex_data)
}
}
\references{
\url{https://realized.oxford-man.ox.ac.uk/}
}
\seealso{
\code{\link{xts}}
}
