% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_tpu}
\alias{get_tpu}
\title{Get Trade Policy Uncertainty (TPU) indices}
\usage{
get_tpu(region = "China")
}
\arguments{
\item{region}{a character indicating the region of the TPU you want. The default
is "China" regions. Can also be "Japan" and "US".}
}
\value{
return an xts data object containing the TPU for the chosen region
}
\description{
Takes in the region name
}
\examples{
china_tpu <- get_tpu()
plot(china_tpu)
# get Japan TPU
jap_tpu <- get_tpu("Japan")
library(dygraphs)
dygraph(jap_tpu)
}
\references{
\url{https://www.policyuncertainty.com/}
}
\seealso{
xts
}
