% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_WUI}
\alias{get_WUI}
\title{Get World Uncertainty Index (WUI) data}
\usage{
get_WUI(type = "F1")
}
\arguments{
\item{type}{sheet option from the official dataset
F1: overall uncertainty across the globe.
F2: trade uncertainty across the globe.
T1: World Uncertainty Index (WUI) at the global level (simple average and GDP weighted average),
income level (advanced, emerging, and low-income economies), and regional level
 (Africa, Asia and the Pacific, Europe, Middle East and Central Asia,
 and Western Hemisphere).
 T2: World Uncertainty Index (WUI) for 143 countries
 T3:  total count of the word "uncertainty" (or its variant) in the EIU
 country reports for 143 countries and from the 1950s
 T4: count of the total number of words in the EIU reports for 143 countries
 and from the 1950s
 T5: count of the total number of pages in the EIU countries reports for
 143 countries from the 1950s
 T6：three-quarter weighted moving average of the World Uncertainty Index (WUI) for
 143 countries from the 1950s
 T8: aggregate World Trade Uncertainty (WTU) index as well as the index by country.}
}
\value{
return an xts data object
}
\description{
Get World Uncertainty Index (WUI) data
}
\details{
The dataset includes the World Uncertainty Index (WUI) at the global level,
as well as by income, region, and country levels.
}
\examples{
\dontrun{
wui_ave <- get_WUI("F1")
plot(wui_ave)
}
}
\references{
\url{https://worlduncertaintyindex.com/data/}
}
\seealso{
\code{\link{xts}}
}
