% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_EMV}
\alias{get_EMV}
\title{Get US Equity Market Volatility Index (EMV) data}
\usage{
get_EMV(all = T)
}
\arguments{
\item{all}{logical, if TRUE return all EMV categories.}
}
\value{
an xts data object
}
\description{
Get US Equity Market Volatility Index (EMV) data
}
\examples{
\dontrun{
## it take about 15 seconds to run this example
emv_data <- get_EMV(all = FALSE)
plot(emv_data)
}
}
\references{
\url{https://www.policyuncertainty.com/EMV_monthly.html}
}
\seealso{
\code{\link{xts}}
}
