% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_FSI}
\alias{get_FSI}
\title{Get Financial Stress Indicator (FSI) data}
\usage{
get_FSI(freq = "monthly")
}
\arguments{
\item{freq}{either "monthly" or "quarterly".}
}
\value{
an xts data object.
}
\description{
Get Financial Stress Indicator (FSI) data
}
\examples{
# for monthly FSI
fsi_mon <- get_FSI()
plot(fsi_mon)
fsi_quar <- get_FSI("quarterly")
plot(fsi_quar)
}
\references{
\url{https://www.policyuncertainty.com/financial_stress.html}
}
\seealso{
\code{\link{xts}}
}
