% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_IRI}
\alias{get_IRI}
\title{Get Immigration Related Index (IRI) data}
\usage{
get_IRI(region = "all")
}
\arguments{
\item{region}{choose from UK, USA, Germany, France for Migrant related EPU and IRI for the specified region. Default returns all regions data.}
}
\value{
an xts data object
}
\description{
Get Immigration Related Index (IRI) data
}
\examples{
usa_iri <- get_IRI("USA")
plot(usa_iri)
}
\references{
\url{https://www.policyuncertainty.com/immigration_fear.html}
}
\seealso{
\code{\link{xts}}
}
