% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_TPU}
\alias{get_TPU}
\title{Get Trade Policy Uncertainty (TPU) data}
\usage{
get_TPU(region = "China")
}
\arguments{
\item{region}{a character indicating the region of the TPU you want. The default
is "China" regions. Can also be "Japan" and "US".}
}
\value{
return an xts data object containing the TPU for the chosen region
}
\description{
Takes in the region name
}
\examples{
china_tpu <- get_TPU()
plot(china_tpu)
# get Japan TPU
jap_tpu <- get_TPU("Japan")
library(dygraphs)
dygraph(jap_tpu)
}
\references{
\url{https://www.policyuncertainty.com/trade_uncertainty.html}
}
\seealso{
\code{\link{xts}}
}
