% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{get_WUI}
\alias{get_WUI}
\title{Get World Uncertainty Index (WUI) data}
\usage{
get_WUI(type = "F1")
}
\arguments{
\item{type}{sheet option from the official excel file.}
}
\value{
an xts data object
}
\description{
Get World Uncertainty Index (WUI) data
}
\examples{
wui_ave <- get_WUI("F1")
plot(wui_ave)
}
\references{
\url{https://worlduncertaintyindex.com/data/}
}
\seealso{
\code{\link{xts}}
}
